+++ noodles-csi/src/index/reference_sequence/builder.rs
@@ -11,6 +11,7 @@ use super::{
 #[derive(Debug)]
 pub struct Builder {
     bin_builders: HashMap<usize, bin::Builder>,
+    linear_index: Vec<Option<bgzf::VirtualPosition>>,
     start_position: bgzf::VirtualPosition,
     end_position: bgzf::VirtualPosition,
     mapped_record_count: u64,
@@ -28,6 +29,7 @@ impl Builder {
         chunk: Chunk,
     ) {
         self.update_bins(min_shift, depth, start, end, chunk);
+        self.update_linear_index(start, end, chunk);
         self.update_metadata(is_mapped, chunk);
     }
 
@@ -58,6 +60,12 @@ impl Builder {
 
         let bins: Vec<_> = self.bin_builders.into_values().map(|b| b.build()).collect();
 
+        let linear_index = self
+            .linear_index
+            .into_iter()
+            .map(|p| p.unwrap_or_default())
+            .collect();
+
         let metadata = Metadata::new(
             self.start_position,
             self.end_position,
@@ -65,7 +73,7 @@ impl Builder {
             self.unmapped_record_count,
         );
 
-        ReferenceSequence::new(bins, Vec::new(), Some(metadata))
+        ReferenceSequence::new(bins, linear_index, Some(metadata))
     }
 
     fn update_bins(
@@ -88,6 +96,21 @@ impl Builder {
         builder.add_chunk(chunk);
     }
 
+    fn update_linear_index(&mut self, start: Position, end: Position, chunk: Chunk) {
+        use super::LINEAR_INDEX_WINDOW_SIZE;
+
+        let linear_index_start_offset = (usize::from(start) - 1) / LINEAR_INDEX_WINDOW_SIZE;
+        let linear_index_end_offset = (usize::from(end) - 1) / LINEAR_INDEX_WINDOW_SIZE;
+
+        if linear_index_end_offset >= self.linear_index.len() {
+            self.linear_index.resize(linear_index_end_offset + 1, None);
+        }
+
+        for i in linear_index_start_offset..=linear_index_end_offset {
+            self.linear_index[i].get_or_insert(chunk.start());
+        }
+    }
+
     fn update_metadata(&mut self, is_mapped: bool, chunk: Chunk) {
         if is_mapped {
             self.mapped_record_count += 1;
@@ -104,6 +127,7 @@ impl Default for Builder {
     fn default() -> Self {
         Self {
             bin_builders: HashMap::new(),
+            linear_index: Vec::new(),
             start_position: bgzf::VirtualPosition::MAX,
             end_position: bgzf::VirtualPosition::MIN,
             mapped_record_count: 0,
@@ -114,6 +138,8 @@ impl Default for Builder {
 
 #[cfg(test)]
 mod tests {
+    use crate::binning_index::ReferenceSequenceExt;
+
     use super::*;
 
     #[test]
@@ -126,48 +152,82 @@ mod tests {
         builder.add_record(
             MIN_SHIFT,
             DEPTH,
-            Position::try_from(2)?,
-            Position::try_from(5)?,
+            Position::try_from(8)?,
+            Position::try_from(13)?,
             true,
             Chunk::new(
-                bgzf::VirtualPosition::from(55),
-                bgzf::VirtualPosition::from(89),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(9),
             ),
         );
 
         builder.add_record(
             MIN_SHIFT,
             DEPTH,
-            Position::try_from(8)?,
-            Position::try_from(13)?,
+            Position::try_from(121393)?,
+            Position::try_from(196418)?,
             false,
             Chunk::new(
-                bgzf::VirtualPosition::from(89),
-                bgzf::VirtualPosition::from(144),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(3473408),
             ),
         );
 
         let actual = builder.build();
 
         let expected = ReferenceSequence::new(
-            vec![Bin::new(
-                4681,
-                bgzf::VirtualPosition::from(55),
-                vec![Chunk::new(
-                    bgzf::VirtualPosition::from(55),
-                    bgzf::VirtualPosition::from(144),
-                )],
-            )],
-            Vec::new(),
+            vec![
+                Bin::new(
+                    4681,
+                    bgzf::VirtualPosition::from(0),
+                    vec![Chunk::new(
+                        bgzf::VirtualPosition::from(0),
+                        bgzf::VirtualPosition::from(9),
+                    )],
+                ),
+                Bin::new(
+                    73,
+                    bgzf::VirtualPosition::from(0),
+                    vec![Chunk::new(
+                        bgzf::VirtualPosition::from(9),
+                        bgzf::VirtualPosition::from(3473408),
+                    )],
+                ),
+            ],
+            vec![
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+                bgzf::VirtualPosition::from(9),
+            ],
             Some(Metadata::new(
-                bgzf::VirtualPosition::from(55),
-                bgzf::VirtualPosition::from(144),
+                bgzf::VirtualPosition::from(0),
+                bgzf::VirtualPosition::from(3473408),
                 1,
                 1,
             )),
         );
 
-        assert_eq!(actual, expected);
+        for expected_bin in expected.bins() {
+            let actual_bin = actual
+                .bins()
+                .iter()
+                .find(|b| b.id() == expected_bin.id())
+                .expect("missing bin");
+
+            assert_eq!(actual_bin, expected_bin);
+        }
+
+        assert_eq!(actual.linear_index(), expected.linear_index());
+        assert_eq!(actual.metadata(), expected.metadata());
 
         Ok(())
     }